Messaging System with RabbitMQ Celery and Python Application behind NGINX/requirements.txt
flask
celery
gunicorn
python-dotenv